feat(blocksync): set the max number of (concurrently) downloaded blocks
to {peersCount * 10}

where 10 is the new max requests per peer.

Previously, the number of requesters (and consequently the number of
requested blocks) was 600 and did not depend on the number of peers.
There's no point in creating more than that since N of requests per peer
is capped and we know the number of peers.

The previous max requests per peer was 20, but it's too big. Given 50
peers, we would request 1000 blocks (50 * 20) ahead of time.