feat(blocksync)!: set the max number of (concurrently) downloaded blocks to {peersCount * 20} (#2467)
where 20 is the new max requests per peer.

Previously, the number of requesters (and consequently the number of
requested blocks) was 600 and did not depend on the number of peers.
There's no point in creating more than that since N of requests per peer
is capped and we know the number of peers.

Closes #2465
Fixes #1283
[api-breaking]

---

#### PR checklist

- [x] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec