fix(e2e): Increase timeout when waiting for new block (#2606)
Some nightly test scenarios are failing because the test that checks
that new blocks are being created timeouts. This happens on larger
networks with many perturbations, on a few heights. Currently the
timeout is 5 seconds, but in some cases it takes the network more than
20 seconds to create a block, so this PR changes it to 30 seconds.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [ ] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec