refactor(consensus): separate sending votes from "pick votes" logic (#2659)
This is done by @adizere , @ebuchman , and @sergio-mena as part of a
refactoring spree :-)

The goal of this is to improve testability of gossipVotesRoutine by
separating out the logic for picking which vote to send. We introduced a
new pure function, pickVoteToSend, which returns a vote, and then call
the new SendVoteSetHasVote method to send it. This greatly simplifies
the loop in gossipVoteRoutine, and allows the pickVoteToSend function to
be independently tested.

Note for reviewers:
* This is pure refactoring, no logic change was intended.
* This PR can be reviewed commit by commit

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [ ] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Daniel <daniel.cason@informal.systems>