fix(mempool): Move senders from reactor back to CList entries (#3131)
Closes #3084 

This PR re-introduces the `TxInfo` parameter with sender information
into the signature of `CheckTx` in the `Mempool` interface. The new
signature is `CheckTx(tx types.Tx, sender p2p.ID) (*abcicli.ReqRes,
error)`, instead of `CheckTx(tx types.Tx) (*abcicli.ReqRes, error)`.

PR #1010 moved the senders from the CList mempool entries to the
reactor. The reason for the changes in this PR is that storing the
sender (in the reactor, through a callback) after storing the
transaction (in the CList struct) introduces a small delay during which
the transaction doesn't have a sender, allowing, in a few cases, the
gossip routines to send back the transaction to the sender. Now the
transaction and the (first) sender are stored together when the
transaction is added to the mempool.

Note that the changes in #1010 didn't reach `v0.38.x`, so this change
will partially revert the mempool API in `main` and `v1.x`.

For reviewers, please see the github comments on the code.

<!--

Please add a reference to the issue that this PR addresses and indicate
which
files are most critical to review. If it fully addresses a particular
issue,
please include "Closes #XXX" (where "XXX" is the issue number).

If this PR is non-trivial/large/complex, please ensure that you have
either
created an issue that the team's had a chance to respond to, or had some
discussion with the team prior to submitting substantial pull requests.
The team
can be reached via GitHub Discussions or the Cosmos Network Discord
server in
the #cometbft channel. GitHub Discussions is preferred over Discord as
it
allows us to keep track of conversations topically.
https://github.com/cometbft/cometbft/discussions

If the work in this PR is not aligned with the team's current
priorities, please
be advised that it may take some time before it is merged - especially
if it has
not yet been discussed with the team.

See the project board for the team's current priorities:
https://github.com/orgs/cometbft/projects/1

-->

---

#### PR checklist

- [X] Tests written/updated
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [X] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [X] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Sergio Mena <sergio@informal.systems>