fix(e2e): add TTL for txs in kvstore to remove txs stuck in mempool (#3201)
The e2e test `TestApp_Tx` is sometimes failing on testnets with full
nodes doing state sync. This is the scenario:
- Node A starts with state sync, receives a snapshot at height H,
switches to consensus mode, and enables the mempool.
- Then node A receives from a peer a transaction that was included in a
block at height lower than H. This transaction will get stuck in the
mempool forever. The reason is that in e2e transactions are always
valid, so it cannot be removed from the mempool during rechecking. It
can only be removed when committing a block that includes the
transaction. But the transaction was already committed at height H
before node A started.

This fix implements in the e2e application a TTL for transactions as a
weak form of replay protection. Real applications should implement
replay protection, so they shouldn't have this problem.

---

#### PR checklist

- [ ] Tests written/updated
- [ ] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [ ] Updated relevant documentation (`docs/` or `spec/`) and code
comments
- [ ] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec