perf(p2p)!: Remove PeerSendBytesTotal metric (#3184)
Component of #2840. As we discussed in the issue, we want to deprecate
the peer send bytes total metric because:
- This is a costly metric (see pprof)
- This increases the cardinality of labels in Prometheus, which can slow
things down in Prometheus significantly.
- Its not really useful at this level of abstraction. We don't have
reported usecases of this being used yet. If we want to do it in the
future, we should more likely architect it as updating a live go
variable, and then sending to our metrics instance on a polling ticker
(ala PeerSendQueue metric). The right levels that are interesting for
outbound traffic could be the consensus gossip routines or blocksync.

To see its costly, 21% of all CPU time for live osmosis nodes in the
latest Osmosis patch release is in peer.Send. over half of that time is
in these metrics:

![image](https://github.com/cometbft/cometbft/assets/6440154/efba6e22-da43-4b97-ad1b-600a13738b5f)

Please advise for how we should go about deprecating / removing this
metric? Feel free to close this PR / say you want to handle it
separately.

---

#### PR checklist

- [X] Tests written/updated - N/A
- [x] Changelog entry added in `.changelog` (we use
[unclog](https://github.com/informalsystems/unclog) to manage our
changelog)
- [x] Updated relevant documentation (`docs/` or `spec/`) and code
comments - I didn't find any relevant docs
- [x] Title follows the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) spec

---------

Co-authored-by: Andy Nogueira <me@andynogueira.dev>
Co-authored-by: Anton Kaliaev <anton.kalyaev@gmail.com>
(cherry picked from commit 80114b2677ad95f45fea2293f2c6fe96d0cee14e)